[frontend] Move dto types to lib/types.ts
