+++ frontend/app/lib/client-auth.tsx
@@ -1,12 +1,6 @@
 "use client";
-import type { UserEntity } from "@/app/lib/types";
+import type { JwtPayload, UserEntity } from "@/app/lib/types";
 import { createContext, useContext } from "react";
-
-export type JwtPayload = {
-  userId: number;
-  isTwoFactorEnabled: boolean;
-  isTwoFactorAuthenticated: boolean;
-};
 
 export type AuthContextType = {
   payload?: JwtPayload;
