--- frontend/app/ui/room/types.ts
@@ -1,10 +1,0 @@
-export type Message = {
-  user: User;
-  content: string;
-};
-
-export type User = {
-  id: number;
-  name: string;
-  avatarURL?: string;
-};
